requirements.txt
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
pymoo==0.6.1.5
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
torch==2.9.0
ultralytics==8.3.225
faiss-cpu==1.12.0
